+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
  
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
+
+## [2.0.0rc1] - 2022-09-15
+
+* Modularized build system
+* Centralized configuration @ /service-config/config
+* Centralized secrets @  /service-config/config/secrets.yaml
+ 
+## [] - 2022-08-29
+
+* Add several containers (home assistant, influxdb, ...)
+* Drop esp-idf container
  
 ## [1.0.1] - 2022-07-10